tests/wpt/tests/html/canvas/element/path-objects/2d.path.isPointInPath.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.isPointInPath.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.isPointInPath.basic</h1>
<p class="desc">Verify the winding rule in isPointInPath works for for rect path.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify the winding rule in isPointInPath works for for rect path.");
_addTest(function(canvas, ctx) {

  canvas.width = 200;
  canvas.height = 200;

  // Testing default isPointInPath
  ctx.beginPath();
  ctx.rect(0, 0, 100, 100);
  ctx.rect(25, 25, 50, 50);
  _assertSame(ctx.isPointInPath(50, 50), true, "ctx.isPointInPath(50, 50)", "true");
  _assertSame(ctx.isPointInPath(NaN, 50), false, "ctx.isPointInPath(NaN, 50)", "false");
  _assertSame(ctx.isPointInPath(50, NaN), false, "ctx.isPointInPath(50, NaN)", "false");

  // Testing nonzero isPointInPath
  ctx.beginPath();
  ctx.rect(0, 0, 100, 100);
  ctx.rect(25, 25, 50, 50);
  _assertSame(ctx.isPointInPath(50, 50, 'nonzero'), true, "ctx.isPointInPath(50, 50, 'nonzero')", "true");

  // Testing evenodd isPointInPath
  ctx.beginPath();
  ctx.rect(0, 0, 100, 100);
  ctx.rect(25, 25, 50, 50);
  _assertSame(ctx.isPointInPath(50, 50, 'evenodd'), false, "ctx.isPointInPath(50, 50, 'evenodd')", "false");

  // Testing extremely large scale
  ctx.save();
  ctx.scale(Number.MAX_VALUE, Number.MAX_VALUE);
  ctx.beginPath();
  ctx.rect(-10, -10, 20, 20);
  _assertSame(ctx.isPointInPath(0, 0, 'nonzero'), true, "ctx.isPointInPath(0, 0, 'nonzero')", "true");
  _assertSame(ctx.isPointInPath(0, 0, 'evenodd'), true, "ctx.isPointInPath(0, 0, 'evenodd')", "true");
  ctx.restore();

  // Check with non-invertible ctm.
  ctx.save();
  ctx.scale(0, 0);
  ctx.beginPath();
  ctx.rect(-10, -10, 20, 20);
  _assertSame(ctx.isPointInPath(0, 0, 'nonzero'), false, "ctx.isPointInPath(0, 0, 'nonzero')", "false");
  _assertSame(ctx.isPointInPath(0, 0, 'evenodd'), false, "ctx.isPointInPath(0, 0, 'evenodd')", "false");
  ctx.restore();

});
</script>
